New Unikraft Homepage Design
This commit introduces a large series of changes which represent the
latest version of the Unikraft OSS homepage and documentation website.

The commit is purposefully created as a larger commit in order to
represent significant change to the existing codebase.  The current
codebase uses the ZDoc template which has now been substantially
modified.

Signed-off-by: Alexander Jung <alex@unikraft.org>