github/linter: Fix sentence end inside backticks
The linter checks for . ? ! outside an escaped sequence.
The backticks set the outside variable as `!outside`, while the
parentheses set it as true or false, depending of the context.
This does not work when backticks are used inside parentheses, making
the text between the backticks be interpreted as unescaped.

i.e. "This sentence (having this `.escaped` sequence)."
gives a linter error, detecting the dot inside backticks
as end of sentence.

This commit changes that, using a different variable for
checking if we are inside backticks.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>