Update community/contacts
This commit updates the primary contacts table of the community page
to reflect the current mail addresses and affiliation of the contacts. The
commit only puts core company members in for now to restrict the
number of contacts also with regard to future growth of the company.
Nevertheless, additional contacts may be added later if needed (also
from the community).

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>