Add SoCC'22 paper
Add material for SoCC'22 conference. SoCC'22 is 13th edition of the annual ACM Symposium on Cloud Computing, the premier conference on cloud computing and is in-person event in San Francisco, CA from November 8th to 10th, 2022. The accepted paper is "Want More Unikernels? Inflate Them!" which is based on Unikraft.
More info: https://acmsocc.org/2022/

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>
Co-Authored-By: Razvan Deaconescu <razvan@rosedu.org>