hackathons/registration: Hide challenges section
Since the challenges section did not have an index file, empty sections
to every challenge would appear in the sidebar menu.
Add a simple index file that marks the challenges as drafts so the empty
sections do not show up anymore.
The challenges will be refered when needed from a hackathon page, like
it is done in the `usoc23/registration-challenges/*.md` files.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>