refactor(hackathons): Migrate and simplify hackathons content
This refactor migrates the hackathon information pages to a higher-level
directory placement to elevate their content.  However, the content for
the hackathon sessions are removed (for now) to simplify the pages.
Right now this is a mess and many tutorials related to running and
building unikraft are out-of-date with regard to using the tooling.
These documents will be reverted in proper placement once a new
structure is determined.

Signed-off-by: Alexander Jung <alex@unikraft.io>