feat: New site design
This new structure is largerly based on the Chakra UI Documentation
website[0].  As a result, the LICENSE is updated to include relevant
copyright notices as it is already compatible with the existing
LICENSE (MIT).  The design is unique, however, and has been modified
to match Unikraft's main color palette and documentation site needs.

[0]: https://github.com/chakra-ui/chakra-ui-docs

Signed-off-by: Alexander Jung <alex@unikraft.io>