Add Loupe ASPLOS'24 paper
Add our Loupe ASPLOS'24 paper. ASPLOS'24 is the 28th edition of the ACM
International Conference on Architectural Support for Programming
Languages and Operating Systems. The paper, "Loupe: Driving the
Development of OS Compatibility Layers", was accepted to appear in
Volume 1. The methods described in this paper were central to the early
development of Unikraft.

More info: https://www.asplos-conference.org/asplos2024/

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>