fix(docs): Replace github tree links with permalinks
If the links just point to the tree, the content they
point to might change over time. These links don't need
the content to be up-to-date. Compared to where they
pointed to in the Unikraft source tree when these links
were initially created, they still point to the right stuff.

Signed-off-by: Thassilo Schulze <thassilo@unikraft.io>